refactor(#58): extract AppHeaderBrand and AppHeaderLanguages
Both layouts compose the same pieces, so the brand and language markup —
and their accessibility details — live in one place each rather than being
duplicated per layout.
